+++ SpeedyTransforms/src/spectral_gradients.jl
@@ -50,7 +50,7 @@ function curl!(
         curl::LowerTriangularArray,
         u::LowerTriangularArray,
         v::LowerTriangularArray,
-        S::SpectralTransform;
+        S::AbstractSpectralTransform;
         flipsign::Bool = false,
         add::Bool = false,
         kwargs...,
@@ -72,7 +72,7 @@ function divergence!(
         div::LowerTriangularArray,
         u::LowerTriangularArray,
         v::LowerTriangularArray,
-        S::SpectralTransform;
+        S::AbstractSpectralTransform;
         flipsign::Bool = false,
         add::Bool = false,
         kwargs...,
@@ -87,10 +87,10 @@ function _divergence!(
         div::LowerTriangularArray,
         u::LowerTriangularArray,
         v::LowerTriangularArray,
-        S::SpectralTransform;
+        S::AbstractSpectralTransform;
         radius = DEFAULT_RADIUS,
     )
-    (; grad_x_vordiv, grad_y_vordiv1, grad_y_vordiv2) = S
+    (; grad_x_vordiv, grad_y_vordiv1, grad_y_vordiv2) = S.gradients
 
     @boundscheck ismatching(S, div) || throw(DimensionMismatch(S, div))
 
@@ -153,7 +153,7 @@ end
 function divergence(
         u::LowerTriangularArray,
         v::LowerTriangularArray,
-        S::SpectralTransform;
+        S::AbstractSpectralTransform;
         kwargs...
     )
     div = similar(u)
@@ -228,7 +228,7 @@ end
 function curl(
         u::LowerTriangularArray,
         v::LowerTriangularArray,
-        S::SpectralTransform;
+        S::AbstractSpectralTransform;
         kwargs...
     )
     vor = similar(u)
@@ -248,10 +248,10 @@ function UV_from_vor!(
         U::LowerTriangularArray,
         V::LowerTriangularArray,
         vor::LowerTriangularArray,
-        S::SpectralTransform;
+        S::AbstractSpectralTransform;
         radius = DEFAULT_RADIUS,
     )
-    (; vordiv_to_uv_x, vordiv_to_uv1, vordiv_to_uv2) = S
+    (; vordiv_to_uv_x, vordiv_to_uv1, vordiv_to_uv2) = S.gradients
     @boundscheck ismatching(S, U) || throw(DimensionMismatch(S, U))
 
     launch!(architecture(U), SpectralWorkOrder, size(U), _UV_from_vor_kernel!, U, V, vor, vor.spectrum.l_indices, vor.spectrum.lmax, vordiv_to_uv_x, vordiv_to_uv1, vordiv_to_uv2)
@@ -313,7 +313,7 @@ function UV_from_vordiv!(
         S::SpectralTransform;
         radius = DEFAULT_RADIUS,
     )
-    (; vordiv_to_uv_x, vordiv_to_uv1, vordiv_to_uv2) = S
+    (; vordiv_to_uv_x, vordiv_to_uv1, vordiv_to_uv2) = S.gradients
     @boundscheck ismatching(S, U) || throw(DimensionMismatch(S, U))
 
     launch!(architecture(U), SpectralWorkOrder, size(U), _UV_from_vordiv_kernel!, U, V, vor, div, vor.spectrum.l_indices, vor.spectrum.lmax, vordiv_to_uv_x, vordiv_to_uv1, vordiv_to_uv2)
@@ -383,18 +383,18 @@ Keyword arguments
 Default is `add=false`, `flipsign=false`, `inverse=false`. These options can be combined.
 """
 function ∇²!(
-        ∇²alms::LowerTriangularArray,   # Output: (inverse) Laplacian of alms
-        alms::LowerTriangularArray,     # Input: spectral coefficients
-        S::SpectralTransform;           # precomputed eigenvalues
-        add::Bool = false,                # add to output array or overwrite
-        flipsign::Bool = false,           # -∇² or ∇²
-        inverse::Bool = false,            # ∇⁻² or ∇²
-        radius = DEFAULT_RADIUS,        # scale with radius if provided, otherwise unit sphere
+        ∇²alms::LowerTriangularArray,       # Output: (inverse) Laplacian of alms
+        alms::LowerTriangularArray,         # Input: spectral coefficients
+        S::AbstractSpectralTransform;       # precomputed eigenvalues
+        add::Bool = false,                  # add to output array or overwrite
+        flipsign::Bool = false,             # -∇² or ∇²
+        inverse::Bool = false,              # ∇⁻² or ∇²
+        radius = DEFAULT_RADIUS,            # scale with radius if provided, otherwise unit sphere
     )
     @boundscheck ismatching(S, ∇²alms) || throw(DimensionMismatch(S, ∇²alms))
 
     # use eigenvalues⁻¹/eigenvalues for ∇⁻²/∇² based but name both eigenvalues
-    eigenvalues = inverse ? S.eigenvalues⁻¹ : S.eigenvalues
+    eigenvalues = inverse ? S.gradients.eigenvalues⁻¹ : S.gradients.eigenvalues
 
     kernel = flipsign ? (add ? (o, a) -> (o - a) : (o, a) -> -a) :
         (add ? (o, a) -> (o + a) : (o, a) -> a)
@@ -426,8 +426,8 @@ Laplace operator ∇² applied to input `alms`, using precomputed eigenvalues fr
 Acts on the unit sphere, i.e. it omits 1/radius^2 scaling unless
 `radius` keyword argument is provided."""
 function ∇²(
-        alms::LowerTriangularArray,     # Input: spectral coefficients
-        S::SpectralTransform;           # precomputed eigenvalues
+        alms::LowerTriangularArray,             # Input: spectral coefficients
+        S::AbstractSpectralTransform;           # precomputed eigenvalues
         kwargs...,
     )
     ∇²alms = similar(alms)
@@ -448,8 +448,8 @@ InverseLaplace operator ∇⁻² applied to input `alms`, using precomputed
 eigenvalues from `S`. Acts on the unit sphere, i.e. it omits radius^2 scaling unless
 `radius` keyword argument is provided."""
 function ∇⁻²(
-        ∇²alms::LowerTriangularArray,   # Input: spectral coefficients
-        S::SpectralTransform;           # precomputed eigenvalues
+        ∇²alms::LowerTriangularArray,           # Input: spectral coefficients
+        S::AbstractSpectralTransform;           # precomputed eigenvalues
         kwargs...,
     )
     alms = similar(∇²alms)
@@ -466,11 +466,11 @@ Acts on the unit sphere, i.e. it omits radius^2 scaling unless
 
 """$(TYPEDSIGNATURES) Calls `∇²!(∇⁻²alms, alms, S; add, flipsign, inverse=true)`."""
 function ∇⁻²!(
-        ∇⁻²alms::LowerTriangularArray,  # Output: inverse Laplacian of alms
-        alms::LowerTriangularArray,     # Input: spectral coefficients
-        S::SpectralTransform;           # precomputed eigenvalues
-        add::Bool = false,              # add to output array or overwrite
-        flipsign::Bool = false,         # -∇⁻² or ∇⁻²
+        ∇⁻²alms::LowerTriangularArray,          # Output: inverse Laplacian of alms
+        alms::LowerTriangularArray,             # Input: spectral coefficients
+        S::AbstractSpectralTransform;           # precomputed eigenvalues
+        add::Bool = false,                      # add to output array or overwrite
+        flipsign::Bool = false,                 # -∇⁻² or ∇⁻²
         kwargs...,
     )
     inverse = true
@@ -481,13 +481,13 @@ end
 in `dpdx` (zonal derivative) and `dpdy` (meridional derivative). The gradient operator acts
 on the unit sphere and therefore omits the 1/radius scaling unless `radius` keyword argument is provided."""
 function ∇!(
-        dpdx::LowerTriangularArray,     # Output: zonal gradient
-        dpdy::LowerTriangularArray,     # Output: meridional gradient
-        p::LowerTriangularArray,        # Input: spectral coefficients
-        S::SpectralTransform;           # includes precomputed arrays
-        radius = DEFAULT_RADIUS,        # scale with radius if provided, otherwise unit sphere
+        dpdx::LowerTriangularArray,             # Output: zonal gradient
+        dpdy::LowerTriangularArray,             # Output: meridional gradient
+        p::LowerTriangularArray,                # Input: spectral coefficients
+        S::AbstractSpectralTransform;           # includes precomputed arrays
+        radius = DEFAULT_RADIUS,                # scale with radius if provided, otherwise unit sphere
     )
-    (; grad_y1, grad_y2) = S
+    (; grad_y1, grad_y2) = S.gradients
     (; m_indices) = p.spectrum
     @boundscheck ismatching(S, p) || throw(DimensionMismatch(S, p))
 
@@ -529,7 +529,7 @@ end
 """$(TYPEDSIGNATURES) The zonal and meridional gradient of `p`
 using an existing `SpectralTransform` `S`. Acts on the unit sphere,
 i.e. it omits 1/radius scaling unless `radius` keyword argument is provided."""
-function ∇(p::LowerTriangularArray, S::SpectralTransform; kwargs...)
+function ∇(p::LowerTriangularArray, S::AbstractSpectralTransform; kwargs...)
     dpdx = similar(p)
     dpdy = similar(p)
     ∇!(dpdx, dpdy, p, S; kwargs...)
@@ -548,7 +548,7 @@ end
 Transform to spectral space, takes the gradient and unscales the 1/coslat
 scaling in the gradient. Acts on the unit-sphere, i.e. it omits 1/radius scaling unless
 `radius` keyword argument is provided. Makes use of an existing spectral transform `S`."""
-function ∇(field::AbstractField, S::SpectralTransform; kwargs...)
+function ∇(field::AbstractField, S::AbstractSpectralTransform; kwargs...)
     p = transform(field, S)
     dpdx, dpdy = ∇(p, S; kwargs...)
     dpdx_grid = transform(dpdx, S, unscale_coslat = true)