+++ SpeedyWeather/src/SpeedyWeather.jl
@@ -99,12 +99,13 @@ export zonal_mean
 using SpeedyTransforms
 
 export SpeedyTransforms, SpectralTransform
+export MatrixSpectralTransform
 export transform, transform!
 export curl, divergence, curl!, divergence!
 export ∇, ∇², ∇⁻², ∇!, ∇²!, ∇⁻²!
 export power_spectrum
 
-import SpeedyTransforms: prettymemory
+import SpeedyTransforms: AbstractSpectralTransform, prettymemory
 
 # to be defined in GeoMakie extension
 export animate, globe